// fe/src/lib/api/generated/model/routineEvaluationPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface RoutineEvaluationPayload {
  academic_adjustment_rating?: number | null;
  emotional_adjustment_rating?: number | null;
  financial_adjustment_rating?: number | null;
  other_adjustment?: string;
  physical_adjustment_rating?: number | null;
  recommendations?: string;
  social_adjustment_rating?: number | null;
  special_concern?: string;
  spiritual_adjustment_rating?: number | null;
}